Separate ObjC and Migration into Own Files
